+++ benchmark/searching/binary/binary_search_benchmark.dart
@@ -9,12 +9,13 @@ void main() {
   final int times = 2000;
   final int size = 100 * 1000 * 1000;
   final comp = ((num a, num b) => (a - b).toInt());
+  final original = List<int>.generate(size, (i) => i);
 
   group("Benchmark binary searching", () {
     group("In a sorted list of $size numbers", () {
       List<int> list = [];
       setUp(() {
-        list = List<int>.generate(size, (i) => i);
+        list = original;
       });
 
       benchmark('collection.binarySearch()', () {
@@ -34,7 +35,7 @@ void main() {
     group("In a sorted list of $size numbers with repeated items", () {
       List<int> list = [];
       setUp(() {
-        list = List<int>.generate(size, (i) => (i * 1000 / size).floor());
+        list = original.map((i) => (i * 1000 / size).floor()).toList();
       });
 
       benchmark('collection.binarySearch()', () {
@@ -54,7 +55,7 @@ void main() {
     group("In a sorted list of $size numbers with a custom comparator", () {
       List<int> list = [];
       setUp(() {
-        list = List<int>.generate(size, (i) => i);
+        list = original;
       });
 
       benchmark('collection.binarySearch()', () {