+++ lib/src/sorting/quicksort_haore.dart
@@ -21,13 +21,13 @@
 /// * [compare] is a custom comparator to order the list elements.
 ///   If it is null and [list] items are not [Comparable], [TypeError] is thrown.
 /// * [threshold] is the maximum limit for which a range can be sorted using
-///   gnome sort.
+///   insertion sort.
 ///
 /// ## Optimizations
 ///
 /// 1. Using iterative approach to avoid recursion. (function calls are slow)
 /// 2. Keeping stack smaller by tail-call optimization. (reduces memory usage)
-/// 3. Use gnome sort on smaller ranges. (configurable by [threshold] parameter)
+/// 3. Use insertion sort on smaller ranges. (configurable by [threshold] parameter)
 /// 4. Exclude items equal to the pivot to avoid worst-case performance on list with
 ///   repeatitive items.
 /// 5. Keeping separate logic for when compare function is provided or not.
@@ -76,13 +76,22 @@ void quickSortHaore<E>(
   if (b + 1 >= e) return;
 
   // add the range to a stack.
+  if (compare == null) {
+    _quickSortDefault(list, b, e, threshold);
+  } else {
+    _quickSortCustom(list, b, e, threshold, compare);
+  }
+}
+
+void _quickSortDefault<E>(List<E> list, int b, int e, int threshold) {
+  E pivot;
+  int l, h, p, q, r;
+
   var stack = <int>[];
   stack.add(b);
   stack.add(e - 1);
 
   // While the stack is not empty
-  E pivot;
-  int l, h, p, q, r;
   while (stack.isNotEmpty) {
     // pop last range from the stack
     h = stack.removeLast();
@@ -93,48 +102,78 @@ void quickSortHaore<E>(
       // Choose the middle item as pivot
       pivot = list[(l + h) >> 1];
 
-      if (compare == null) {
-        // Optimization: if the range is short, sort using gnome sort and break
-        if (l + threshold > h) {
-          _gnomeSortDefault(list, l, h);
-          break;
-        }
+      // Optimization: if the range is short, sort using insertion sort and break
+      if (l + threshold > h) {
+        _insertionSortDefault(list, l, h);
+        break;
+      }
 
-        // Partition list around pivot index
-        p = _partitionRangeDefault(list, l, h);
+      // Partition list around pivot index
+      p = _partitionRangeDefault(list, l, h);
 
-        // Optimization for repeated values: exclude items equal to pivot
-        // exclude the items from the left range that are equal to the pivot
-        q = p;
-        while (l < q && list[q] == pivot) {
-          q--;
-        }
-        // exclude the items from the right range that are equal to the pivot
-        r = p + 1;
-        while (r < h && list[r] == pivot) {
-          r++;
-        }
-      } else {
-        // Optimization: if the range is short, sort using gnome sort and break
-        if (l + threshold > h) {
-          _gnomeSortCustom(list, l, h, compare);
-          break;
-        }
+      // Optimization for repeated values: exclude items equal to pivot
+      // exclude the items from the left range that are equal to the pivot
+      q = p;
+      while (l < q && list[q] == pivot) {
+        q--;
+      }
+      // exclude the items from the right range that are equal to the pivot
+      r = p + 1;
+      while (r < h && list[r] == pivot) {
+        r++;
+      }
 
-        // Partition list around pivot index
-        p = _partitionRangeCustom(list, l, h, compare);
+      // take the left partition if it has more than one element
+      if (l < q) {
+        stack.add(l);
+        stack.add(q);
+      }
 
-        // Optimization for repeated values: exclude items equal to pivot
-        // exclude the items from the left range that are equal to the pivot
-        q = p;
-        while (l < q && compare(list[q], pivot) == 0) {
-          q--;
-        }
-        // exclude the items from the right range that are equal to the pivot
-        r = p + 1;
-        while (r < h && compare(list[r], pivot) == 0) {
-          r++;
-        }
+      // select the right partition to continue the inner loop
+      l = r;
+    }
+  }
+}
+
+void _quickSortCustom<E>(
+    List<E> list, int b, int e, int threshold, Comparator<E> compare) {
+  E pivot;
+  int l, h, p, q, r;
+
+  var stack = <int>[];
+  stack.add(b);
+  stack.add(e - 1);
+
+  // While the stack is not empty
+  while (stack.isNotEmpty) {
+    // pop last range from the stack
+    h = stack.removeLast();
+    l = stack.removeLast();
+
+    // To keep the stack smaller by simulating the tail-call optimization
+    while (l < h) {
+      // Choose the middle item as pivot
+      pivot = list[(l + h) >> 1];
+
+      // Optimization: if the range is short, sort using insertion sort and break
+      if (l + threshold > h) {
+        _insertionSortCustom(list, l, h, compare);
+        break;
+      }
+
+      // Partition list around pivot index
+      p = _partitionRangeCustom(list, l, h, compare);
+
+      // Optimization for repeated values: exclude items equal to pivot
+      // exclude the items from the left range that are equal to the pivot
+      q = p;
+      while (l < q && compare(list[q], pivot) == 0) {
+        q--;
+      }
+      // exclude the items from the right range that are equal to the pivot
+      r = p + 1;
+      while (r < h && compare(list[r], pivot) == 0) {
+        r++;
       }
 
       // take the left partition if it has more than one element
@@ -150,42 +189,31 @@ void quickSortHaore<E>(
 }
 
 /// compare items with default comparision
-void _gnomeSortDefault<E>(List<E> list, int l, int h) {
-  E t;
-  int i = l + 1;
-  while (i <= h) {
-    if (i == l) {
-      i++;
-      continue;
+void _insertionSortDefault<E>(List<E> list, int b, int e) {
+  E el;
+  int i, j;
+  for (i = b + 1; i <= e; ++i) {
+    for (j = i; j > b; --j) {
+      if ((list[j - 1] as Comparable).compareTo(list[j]) <= 0) break;
+      el = list[j - 1];
+      list[j - 1] = list[j];
+      list[j] = el;
     }
-    if ((list[i - 1] as Comparable).compareTo(list[i]) <= 0) {
-      i++;
-      continue;
-    }
-    t = list[i - 1];
-    list[i - 1] = list[i];
-    list[i] = t;
-    i--;
   }
 }
 
 /// compare items with custom comparator (slower)
-void _gnomeSortCustom<E>(List<E> list, int l, int h, Comparator<E> compare) {
-  E t;
-  int i = l + 1;
-  while (i <= h) {
-    if (i == l) {
-      i++;
-      continue;
+void _insertionSortCustom<E>(
+    List<E> list, int b, int e, Comparator<E> compare) {
+  E el;
+  int i, j;
+  for (i = b + 1; i <= e; ++i) {
+    for (j = i; j > b; --j) {
+      if (compare(list[j - 1], list[j]) <= 0) break;
+      el = list[j - 1];
+      list[j - 1] = list[j];
+      list[j] = el;
     }
-    if (compare(list[i - 1], list[i]) <= 0) {
-      i++;
-      continue;
-    }
-    t = list[i - 1];
-    list[i - 1] = list[i];
-    list[i] = t;
-    i--;
   }
 }