+++ lib/src/sorting/quicksort_lomuto.dart
@@ -21,13 +21,13 @@
 /// * [compare] is a custom comparator to order the list elements.
 ///   If it is null and [list] items are not [Comparable], [TypeError] is thrown.
 /// * [threshold] is the maximum limit for which a range can be sorted using
-///   gnome sort.
+///   insertion sort.
 ///
 /// ## Optimizations
 ///
 /// 1. Using iterative approach to avoid recursion. (function calls are slow)
 /// 2. Keeping stack smaller by tail-call optimization. (reduces memory usage)
-/// 3. Use gnome sort on smaller ranges. (configurable by [threshold] parameter)
+/// 3. Use insertion sort on smaller ranges. (configurable by [threshold] parameter)
 /// 4. Following Sedwick's optimization and using median-of-3 to choose the pivot.
 ///   (avoiding worst-case performance on already sorted list)
 /// 5. Exclude items equal to the pivot to avoid worst-case performance on list with
@@ -77,13 +77,22 @@ void quickSortLomuto<E>(
   // If the range has less than two items, returns immediately.
   if (b + 1 >= e) return;
 
+  if (compare == null) {
+    _quickSortDefault(list, b, e, threshold);
+  } else {
+    _quickSortCustom(list, b, e, threshold, compare);
+  }
+}
+
+void _quickSortDefault<E>(List<E> list, int b, int e, int threshold) {
+  int l, h, p, q;
+
   // add the range to a stack.
   var stack = <int>[];
   stack.add(b);
   stack.add(e - 1);
 
   // While the stack is not empty
-  int l, h, p, q;
   while (stack.isNotEmpty) {
     // pop last range from the stack
     h = stack.removeLast();
@@ -91,44 +100,71 @@ void quickSortLomuto<E>(
 
     // To keep the stack smaller by simulating the tail-call optimization
     while (l < h) {
-      if (compare == null) {
-        // If there are few elements, sort them using gnome sort and break the loop
-        if (l + threshold > h) {
-          _gnomeSortDefault(list, l, h);
-          break;
-        }
+      // If there are few elements, sort them using insertion sort and break the loop
+      if (l + threshold > h) {
+        _insertionSortDefault(list, l, h);
+        break;
+      }
 
-        // To avoid worst-case pitfall of already sorted list
-        _sortByMedianDefault(list, l, h);
+      // To avoid worst-case pitfall of already sorted list
+      _sortByMedianDefault(list, l, h);
 
-        // Partition list around pivot index
-        p = _partitionRangeDefault(list, l, h);
+      // Partition list around pivot index
+      p = _partitionRangeDefault(list, l, h);
 
-        // Optimization for repeated values on left partition.
-        // exclude the items from the left range that are equal to the pivot
-        q = p + 1;
-        while (q < h && list[q] == list[p]) {
-          q++;
-        }
-      } else {
-        // If there are few elements, sort them using gnome sort and break the loop
-        if (l + threshold > h) {
-          _gnomeSortCustom(list, l, h, compare);
-          break;
-        }
+      // Optimization for repeated values on left partition.
+      // exclude the items from the left range that are equal to the pivot
+      q = p + 1;
+      while (q < h && list[q] == list[p]) {
+        q++;
+      }
 
-        // To avoid worst-case pitfall of already sorted list
-        _sortByMedianCustom(list, l, h, compare);
+      // take the left partition if it has more than one element
+      if (l < p - 1) {
+        stack.add(l);
+        stack.add(p - 1);
+      }
 
-        // Partition list around pivot index
-        p = _partitionRangeCustom(list, l, h, compare);
+      // select the right partition to continue the inner loop
+      l = q;
+    }
+  }
+}
 
-        // Optimization for repeated values on left partition.
-        // exclude the items from the left range that are equal to the pivot
-        q = p + 1;
-        while (q < h && compare(list[q], list[p]) == 0) {
-          q++;
-        }
+void _quickSortCustom<E>(
+    List<E> list, int b, int e, int threshold, Comparator<E> compare) {
+  int l, h, p, q;
+
+  // add the range to a stack.
+  var stack = <int>[];
+  stack.add(b);
+  stack.add(e - 1);
+
+  // While the stack is not empty
+  while (stack.isNotEmpty) {
+    // pop last range from the stack
+    h = stack.removeLast();
+    l = stack.removeLast();
+
+    // To keep the stack smaller by simulating the tail-call optimization
+    while (l < h) {
+      // If there are few elements, sort them using insertion sort and break the loop
+      if (l + threshold > h) {
+        _insertionSortCustom(list, l, h, compare);
+        break;
+      }
+
+      // To avoid worst-case pitfall of already sorted list
+      _sortByMedianCustom(list, l, h, compare);
+
+      // Partition list around pivot index
+      p = _partitionRangeCustom(list, l, h, compare);
+
+      // Optimization for repeated values on left partition.
+      // exclude the items from the left range that are equal to the pivot
+      q = p + 1;
+      while (q < h && compare(list[q], list[p]) == 0) {
+        q++;
       }
 
       // take the left partition if it has more than one element
@@ -144,42 +180,31 @@ void quickSortLomuto<E>(
 }
 
 /// compare items with default comparision
-void _gnomeSortDefault<E>(List<E> list, int l, int h) {
-  E tmp;
-  int i = l + 1;
-  while (i <= h) {
-    if (i == l) {
-      i++;
-      continue;
+void _insertionSortDefault<E>(List<E> list, int b, int e) {
+  E el;
+  int i, j;
+  for (i = b + 1; i <= e; ++i) {
+    for (j = i; j > b; --j) {
+      if ((list[j - 1] as Comparable).compareTo(list[j]) <= 0) break;
+      el = list[j - 1];
+      list[j - 1] = list[j];
+      list[j] = el;
     }
-    if ((list[i - 1] as Comparable).compareTo(list[i]) <= 0) {
-      i++;
-      continue;
-    }
-    tmp = list[i - 1];
-    list[i - 1] = list[i];
-    list[i] = tmp;
-    i--;
   }
 }
 
 /// compare items with custom comparator (slower)
-void _gnomeSortCustom<E>(List<E> list, int l, int h, Comparator<E> compare) {
-  E tmp;
-  int i = l + 1;
-  while (i <= h) {
-    if (i == l) {
-      i++;
-      continue;
+void _insertionSortCustom<E>(
+    List<E> list, int b, int e, Comparator<E> compare) {
+  E el;
+  int i, j;
+  for (i = b + 1; i <= e; ++i) {
+    for (j = i; j > b; --j) {
+      if (compare(list[j - 1], list[j]) <= 0) break;
+      el = list[j - 1];
+      list[j - 1] = list[j];
+      list[j] = el;
     }
-    if (compare(list[i - 1], list[i]) <= 0) {
-      i++;
-      continue;
-    }
-    tmp = list[i - 1];
-    list[i - 1] = list[i];
-    list[i] = tmp;
-    i--;
   }
 }